fix: use release Kustomization for Gatekeeper
A dedicated Kustomization gatekeeper-release is created that the other
two Kustomizations can depend on other than depending on a
Kustomization that isn't maintained as part of this repository. This
makes the whole gatekeeper service self-contained and less prone to
bugs.

This follows the same pattern we use for kubefed so we don't use
different approaches within this repo.