bug: move Knative to "general" category
Having Knative in the foundational category leads to the UI issuing a
warning about not all foundational components being installed and thus
the UI blocks any deployment of additional applications.

As long as we haven't settled on the right category for Knative, I'm
just moving it to "general" for now to enable users to deploy apps
again.