feat: Bump istio (#202)
* feat: Bump istio to 1.11.5

* fix: Remove istio-ca

https://jira.d2iq.com/browse/D2IQ-80630
Previously the istio chart was hardcoding the clusterIssuer
name in the istio-ca cert. This has been fixed in the chart
so that we can pass in the clusterIssuer name dynamically
through the values.

* fix: Remove istio-namespace

https://jira.d2iq.com/browse/D2IQ-80630
Now that we have removed istio-ca cert manifest, we
do not need to also create the istio-system namespace
as a dependency.