feat: sops encryption config params in kommander-vars cm
CLI generates git credentials and sops age secret to encrypt those
credentials during install time. After these are created, kommander
controllers shall be responsible to ensure the secrets are kept
up-to-date (such as when the Certificate information changes) and this
change ensures that the necessary flags are set on kommander controller
deployments.