feat: external-dns type internal -> platform
Updated the type on external-dns from internal to platform. With custom
domain support external-dns should now be made available for all
clusters so it can be configured for any cluster that uses a custom
domain.

Added an icon to the metadata now that the app will be displayed in the
UI.

- Closes D2IQ-90556