fix: Remove koperator from base kustomization
This change was applying automatically to ALL attached clusters during
mgmt cluster upgrade (upgrade of k-apps repo) which had their
kustomization still pointing to the ./common/base. In order to avoid
this, we need to move koperator path out of ./common/base resources and
only install it on the mgmt cluster during install and upgrade.