fix: Run informer in deployment for velero post-install
Create a deployment to run an informer watching velero
BackupStorageLocation objects to update the s3Url in the config to point
to the publicly exposed endpoint so that the "velero backup logs"
command succeeds. This replaces the job that would have only run the
update once at installation -- subsequent upgrades would have reset the
endpoint back to the default internal svc address.