fix: increase failureThreshold of DKA liveness probe #1407
This PR increases to 6 the failureThreshold for the DKA liveness
probe to prevent failures when the number of clusters to parse is >=50.
The longer time required to read the configuration file made the
application to not respond correctly to liveness probe,
resulting in CrashLoopBackoff.
Ref. https://d2iq.atlassian.net/browse/D2IQ-94340