feat: Set kubetunnel priority to system-cluster-critical
DKP priority classes are created after attachment of clusters. However,
for tunneled clusters, the kubetunnel agent components must be deployed
on the cluster first before getting the kubeconfig to access the
cluster. Therefore, we cannot use a dkp priority class for the
kubetunnel agent, since it will not exist on the cluster prior to the
attachment.