chore: bump kubetunnel to 0.0.21 (#1283)
* chore: Update kubetunnel 0.0.15 => 0.0.16

* chore: Move kubetunnel/0.0.15 => kubetunnel/0.0.16

* feat: Update licenses with updated and new kubetunnel images

* feat: Bump kubetunnel to 0.0.17

* feat: Bump kubetunnel to v0.0.18

* feat: dkp critical for kubetunnel

* feat: Bump kubetunnel to 0.0.19

* fix: Remove kubetunnel reverse-proxy image

The reverse-proxy component is no longer deployed with the kubetunnel
chart and has been moved into its own chart.

* feat: Bump kubetunnel to 0.0.20

* feat: Bump kubetunnel to 0.0.21

* feat: Set kubetunnel priority to system-cluster-critical

DKP priority classes are created after attachment of clusters. However,
for tunneled clusters, the kubetunnel agent components must be deployed
on the cluster first before getting the kubeconfig to access the
cluster. Therefore, we cannot use a dkp priority class for the
kubetunnel agent, since it will not exist on the cluster prior to the
attachment.

---------

Co-authored-by: d2iq-mergebot <ci-mergebot@d2iq.com>
Co-authored-by: Grace Do <xgrace@gmail.com>
Co-authored-by: York Chen <ychen@d2iq.com>