ci: add update-licenses label with Mesosphere bot
According to this page [1]

> When you use the repository's `GITHUB_TOKEN` to perform tasks, events
triggered by the `GITHUB_TOKEN`, with the exception of `workflow_dispatch`
and `repository_dispatch`, will not create a new workflow run. This
prevents you from accidentally creating recursive workflow runs.

I believe that's the reason we need to manually remote and re-add the
`update-licenses` label for the automatic update to trigger.

1 - https://docs.github.com/en/actions/security-guides/automatic-token-authentication#using-the-github_token-in-a-workflow