ci: update licenses with user token
According to this page [1]

> When you use the repository's `GITHUB_TOKEN` to perform tasks, events
triggered by the `GITHUB_TOKEN`, with the exception of `workflow_dispatch`
and `repository_dispatch`, will not create a new workflow run. This
prevents you from accidentally creating recursive workflow runs.

I believe that's the reason we need to manually close and re-open the PR
to trigger a verification on the updated head with the automated license
change commit.

1 - https://docs.github.com/en/actions/security-guides/automatic-token-authentication#using-the-github_token-in-a-workflow