fix(dka): Change default healthCheckPath to fix k8s probes
Previously the paths used by probes returned large responses when the
number of clusters was large (>60ish). This seems to cause the kubelet
to restart DKA pod. Changing the healthCheckPath to a static path
ensures smaller responses and is safe to do because the http server is
only started after full initialization of managed cluster endpoints.